Fold object handling into session module
Move BrokerSession object/reference methods and private object registry types into session.rs, then remove the now-empty object module.
